move LazyBlob in utils and dont export it
